Add files via upload
Capabilities: able to take input on the canvas, resize the image, and feed it into the neural network. Classifies the number with mediocre accuracy with major discrepancy when the digit drawn by users is 9. 

Improvements to be made: improve the accuracy of the model. Try to find a dataset to train the model on that has bigger image inputs? Expand the model to be able to classify multi-digit numbers.